% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.rtests-package.R
\docType{package}
\name{inc.rtests-package}
\alias{inc.rtests}
\alias{inc.rtests-package}
\title{inc.rtests: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ixpantia.github.io/inc.rtests/}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

}
\keyword{internal}
