% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Create a linear model formula}
\usage{
linear_model(formula)
}
\arguments{
\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}
}
\description{
Create a linear model formula
}
\examples{
x <- 1:3
y <- 2 * x

linear_model(y ~ x)
}
