% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtests-package.R
\docType{package}
\name{rtests-package}
\alias{rtests}
\alias{rtests-package}
\title{rtests: Testing Demo}
\description{
Show by example how to test R code with testthat, with a focus on testing models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dsincubator/rtests}
  \item \url{https://dsincubator.github.io/rtests/}
  \item Report bugs at \url{https://github.com/dsincubator/rtests/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

}
\keyword{internal}
